Add a Timeout option on OdooClient
Note that this is not currently working properly if you have multiple OdooClient instances with different timeout.
This is because OdooClient share a single static instance of HttpClient.

I disagree with the fact that OdooClient have a Singleton of HttpClient but didn't want to confuse everyone with a complicated merge request.